Refactor password strength check
